Create the main express loaders file with an index.js file to consume it. All this logic has been moved out of app.js, and bin/www removed.
Update app.js to consume config and loaders to make it a lot cleaner.
