--- bin/www
@@ -1,17 +1,0 @@
-// This file is used to launch the
-// Express server and sets up the listener
-
-const app = require('../app')
-
-// Configure the app to listen for incoming
-// traffic on HOST and PORT
-const HOST = process.env.HOST || 'localhost'
-const PORT = process.env.PORT || 5000
-
-
-//-------------------------
-// set up server listening
-//-------------------------
-app.listen( PORT,
-            () => console.log(`Listening on http://${HOST}:${PORT}`)
-)
