+++ loaders/index.js
@@ -1,0 +1,10 @@
+const expressLoader = require('./express');
+
+module.exports = {
+    init: async ({ expressApp }) => {
+
+        await expressLoader.init( expressApp );
+        console.log('Express Intialized');
+    }
+
+}
